# backend/artifacts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clinimetria",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/sorcerer/recommend": {
      "post": {
        "tags": [
          "sorcerer"
        ],
        "summary": "Recommend",
        "operationId": "recommend_api_v1_sorcerer_recommend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SorcererRecommendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SorcererRecommendation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sorcerer/apply": {
      "post": {
        "tags": [
          "sorcerer"
        ],
        "summary": "Apply",
        "operationId": "apply_api_v1_sorcerer_apply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SorcererApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Apply Api V1 Sorcerer Apply Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SorcererApplyRequest": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "recommendation": {
            "type": "object",
            "title": "Recommendation"
          },
          "variables": {
            "type": "object",
            "title": "Variables"
          },
          "test_config": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Config"
          },
          "alpha": {
            "type": "number",
            "title": "Alpha",
            "default": 0.05
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "recommendation",
          "variables"
        ],
        "title": "SorcererApplyRequest"
      },
      "SorcererRecommendRequest": {
        "properties": {
          "goal": {
            "type": "string",
            "enum": [
              "compare_groups",
              "compare_timepoints",
              "relationship",
              "survival",
              "prediction"
            ],
            "title": "Goal",
            "description": "High-level research goal"
          },
          "structure": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "independent",
                  "paired"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Structure",
            "description": "Group structure for compare_groups"
          },
          "data_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "numeric",
                  "categorical"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Type",
            "description": "Outcome data type"
          },
          "groups": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "2",
                  ">2"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Groups",
            "description": "Group count bucket"
          },
          "normal_distribution": {
            "type": "boolean",
            "title": "Normal Distribution",
            "description": "User assumption about normality for numeric outcomes",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "goal"
        ],
        "title": "SorcererRecommendRequest"
      },
      "SorcererRecommendation": {
        "properties": {
          "method_id": {
            "type": "string",
            "title": "Method Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions"
          }
        },
        "type": "object",
        "required": [
          "method_id",
          "name",
          "description"
        ],
        "title": "SorcererRecommendation"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}